+++ app/package.json
@@ -47,6 +47,7 @@
     "@coremedia/ckeditor5-coremedia-studio-integration-mock": "^14.0.2",
     "@coremedia/ckeditor5-dataprocessor-support": "^14.0.2",
     "@coremedia/ckeditor5-font-mapper": "^14.0.2",
+    "@coremedia/ckeditor5-link-common": "^14.0.2",
     "@coremedia/service-agent": "^1.1.5",
     "xml-formatter": "^2.6.1"
   },
