+++ app/src/ckeditor.ts
@@ -64,8 +64,8 @@ import { FilterRuleSetConfiguration } from "@coremedia/ckeditor5-dataprocessor-s
 import { replaceByElementAndClassBackAndForth } from "@coremedia/ckeditor5-coremedia-richtext/compatibility/v10/rules/ReplaceBy";
 import { getHashParam } from "./HashParams";
 import { COREMEDIA_LINK_CONFIG_KEY } from "@coremedia/ckeditor5-coremedia-link/contentlink/LinkBalloonConfig";
-import { LinkAttributesConfig } from "@coremedia/ckeditor5-coremedia-link/link/LinkAttributesConfig";
-import { LinkAttributes } from "@coremedia/ckeditor5-coremedia-link/link/LinkAttributes";
+import { LinkAttributesConfig } from "@coremedia/ckeditor5-link-common/LinkAttributesConfig";
+import { LinkAttributes } from "@coremedia/ckeditor5-link-common/LinkAttributes";
 
 const {
   //@ts-expect-error We have no way to extend icon typing, yet.
