+++ packages/ckeditor5-coremedia-link/src/index-doc.ts
@@ -8,5 +8,4 @@ export * as utils from "./utils";
 
 export * as contentlink from "./contentlink/index-doc";
 export * as lang from "./lang/index-doc";
-export * as link from "./link/index-doc";
 export * as linktarget from "./linktarget/index-doc";
