+++ packages/ckeditor5-link-common/.eslintrc.js
@@ -1,0 +1,12 @@
+module.exports = {
+  parser: "@typescript-eslint/parser",
+  overrides: [
+    {
+      files: ["**/*.ts", "**/*.tsx"],
+      parserOptions: {
+        tsconfigRootDir: __dirname,
+        project: "./src/tsconfig.json",
+      },
+    },
+  ],
+};
