+++ packages/ckeditor5-link-common/README.md
@@ -1,0 +1,4 @@
+# CKEditor 5 Core Common
+
+This package contains some assistive tooling around `@ckeditor/ckeditor5-core`
+API to ease implementation of plugins.
