+++ packages/ckeditor5-link-common/src/Constants.ts
