+++ packages/ckeditor5-link-common/src/FocusUtils.ts
