+++ packages/ckeditor5-link-common/src/LinkAttributes.ts
