+++ packages/ckeditor5-link-common/src/LinkAttributesConfig.ts
