+++ packages/ckeditor5-link-common/src/LinkCleanup.ts
