+++ packages/ckeditor5-link-common/src/RegisterAttributeConfig.ts
