+++ packages/ckeditor5-link-common/src/tsconfig.json
@@ -1,0 +1,33 @@
+{
+  "compilerOptions": {
+    "baseUrl": ".",
+    "rootDir": ".",
+    "outDir": "../dist",
+    "emitDeclarationOnly": true,
+    "declaration": true,
+    "declarationMap": true,
+    "paths": {
+      "@coremedia/ckeditor5-common/*": [
+        "../../ckeditor5-common/src/*"
+      ],
+      "@coremedia/ckeditor5-core-common/*": [
+        "../../ckeditor5-core-common/src/*"
+      ],
+      "@coremedia/ckeditor5-logging/*": [
+        "../../ckeditor5-logging/src/*"
+      ]
+    }
+  },
+  "references": [
+    {
+      "path": "../../ckeditor5-common/src"
+    },
+    {
+      "path": "../../ckeditor5-core-common/src"
+    },
+    {
+      "path": "../../ckeditor5-logging/src"
+    }
+  ],
+  "extends": "../../../tsconfig.json"
+}
